Refactor so std::string and dynd::string are supported
